pyproject.toml
[project]
name = "fitness-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.22.0",
    "asyncio>=4.0.0",
    "asyncpg>=0.30.0",
    "magic-filter>=1.0.12",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy[asyncio]>=2.0.43",
    "sqlmodel>=0.0.25",
    "aiogram-timepicker",
    "pytz>=2025.2",
    "tzwhere>=3.0.3",
    "requests>=2.32.5",
    "fastapi[standard]>=0.119.0",
]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "UP",
    "B",
    "SIM",
    "I"
]

[tool.mypy]
pretty = true
ignore_missing_imports = true

[tool.pyright]
reportInvalidStringEscapeSequence = "none"

[tool.uv.sources]
aiogram-timepicker = { workspace = true }

[tool.uv.workspace]
members = [
    "packages/aiogram-timepicker",
    "packages/timezone-parser",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.8.0",
]
